requirements.txt
pycryptodome>=3.14.1
type-utils>=1.0.2